Update README.md
lab-2  does not enter "$WORK_DIR" directory. the student should be in this directory when done with lab-1, but earlier in this lab-2 document it's described to do "cat << 'EOF' >$WORK_DIR/ping.yml" - which contradicts that the student is already in the work dir. therefore, add cd workdir to be sure that the student is in the right spot.